requirements.txt
fastapi
uvicorn
transformers
torch
pillow
prometheus_client
